Created shared module
moved logout and create password to shared module
